+++ includes/SocketStream.hpp
